line

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,17 +1,18 @@
 #include "search_algos.h"
+#include <stddef.h>
 #include <stdio.h>
 
 /**
  * linear_search - Searches for a value in an array of integers
- *                  using the Linear search algorithm.
+ *                 using the Linear search algorithm.
  * @array: Pointer to the first element of the array to search in.
  * @size: Number of elements in @array.
  * @value: Value to search for.
  *
- * Return: The first index where @value is located,
+ * Return: The index of the first occurrence of @value in @array,
  *         or -1 if @value is not present in @array or if @array is NULL.
  */
-int linear_search(int *array, size_t size, int value)
+ssize_t linear_search(int *array, size_t size, int value)
 {
     if (array == NULL)
         return -1;
